\docType{package}
\name{lcollado754-package}
\alias{lcollado754-package}
\title{Final project for 140.754 (2013 version) by Leonardo Collado-Torres}
\description{
  Analysis of 2,500 peer-to-peer loans issued through the
  Lending Club for the 140.754 final project. The
  instructions are available here:
  https://dl.dropboxusercontent.com/u/7710864/jhsph753/finalProject/finalProject.html
}
\author{
  Leonardo Collado-Torres <lcollado@jhsph.edu>
}
\keyword{package}

