\docType{data}
\name{loansData}
\alias{loansData}
\title{Loans data}
\format{A dataframe with 2,500 rows and 14 columns.
\describe{
\item{Amount.Requested }{ The amount (in dollars) requested in the loan application.}
\item{Amount.Funded.By.Investors }{ The amount (in dollars) loaned to the individual.}
\item{Interest.Rate }{ The lending interest rate.}
\item{Loan.Length }{ The length of time (in months) of the loan.}
\item{Loan.Purpose }{ The purpose of the loan as stated by the applicant.}
\item{Debt.To.Income.Ratio }{ The percentage of consumer's gross income that goes toward paying debts (http://en.wikipedia.org/wiki/Debt-to-income_ratio).}
\item{State }{ The abbreviation for the U.S. state of residence of the loan applicant (http://www.50states.com/abbreviations.htm#.UQ8hxFp2PKo).}
\item{Home.Ownership }{ A variable indicating whether the applicant owns, rents, or has a mortgage on their home.}
\item{Monthly.Income }{ The monthly income of the applicant (in dollars).}
\item{FICO.Range }{ A range indicating the applicants FICO score. This is a measure of the creditworthiness of the applicant  (http://en.wikipedia.org/wiki/Credit_score_in_the_United_States).}
\item{Open.CREDIT.Lines }{ The number of open lines of credit the applicant had at the time of application.}
\item{Revolving.CREDIT.Balance }{ The total amount outstanding all lines of credit (http://www.ehow.com/about_7550001_revolving-credit-balance.html).}
\item{Inquiries.in.the.Last.6.Months }{ When a person applies for credit, they authorize the lender to "inquire" about their creditworthiness. This is the number of such authorized queries in the 6 months before the loan was issued  (http://www.myfico.com/crediteducation/creditinquiries.aspx).}
\item{Employment.Length }{ Length of time employed at current job.}
}}
\description{
  2,500 peer-to-peer loans issued through the Lending Club.
}
\seealso{
  \link{loansData2}
}
\keyword{datasets}

