\docType{data}
\name{states}
\alias{states}
\title{States codes from AQS EPA GOV}
\format{A dataframe with 3,278 rows and 6 columns.
\describe{
\item{State.Name}{ State name.}
\item{State.Abbr}{ Two letter state abbreviation.}
\item{State.Code}{ State code.}
\item{Region}{ Coded 1 to 10 and 25. East is 1 to 5, West is 6 to 10. 25 is outside the US.}
\item{County.Name}{ County name.}
\item{County.Code}{ County code. Can be repeated across different states.}
}}
\description{
  States codes from
  https://aqs.epa.gov/aqsweb/codes/data/StateCountyCodes.csv
}
\keyword{datasets}

